Add support for post commit APIs to rb-gateway
This is to add support for the post commit UI that Review Board currently has.
In order for Review Board Gateway hosted repositories to support this UI,
rb-gateway must provide APIs for getting branches, getting commits, and getting
an individual commit.

The URL for get branches is located at /branches and it requires a repository
as the parameter. The returned JSON looks as follows:

    [
       {
         "name": master,
         "id": "1b6f00c0fe975dd12251431ed2ea561e0acc6d44"
       }
    ]

The URL for get commits is located at /commits and it requries a repository as
the parameter, as well as the branch id. An optional start commit parameter can
be provided to retrieve the commits beginning at the start commit. The returned
JSON looks as follows:

    [
       {
         "author": "John Doe",
         "id": "1b6f00c0fe975dd12251431ed2ea561e0acc6d44",
         "date": "2015-06-27T05:51:39-07:00",
         "message": "Add README.md",
         "parent_id": "bfdde95432b3af879af969bd2377dc3e55ee46e6"
       }
    ]

The URL for get commit is located at /commit and it requries a repository as
the parameter, as well as the commit id. The returned JSON looks as follows:

    {
         "author": "John Doe",
         "id": "1b6f00c0fe975dd12251431ed2ea561e0acc6d44",
         "date": "2015-06-27T05:51:39-07:00",
         "message": "Add README.md",
         "parent_id": "bfdde95432b3af879af969bd2377dc3e55ee46e6",
         "diff": "diff --git a/test b/test
                   index e69de29bb2d1d6434b8b29ae775ad8c2e48c5391..044f599c9a720fe1a7d02e694a8dab492cbda8f0 100644
                   --- a/test
                   +++ b/test
                   @@ -1 +1,3 @@
                   test
                   +
                   +test"
    }

Testing Done:

Verified the responses from the following URLs, run through my browser and self
rb-gateway with a git repository test environment.

- http://localhost:8002/branches?repo=testrepo
- http://localhost:8002/commits?repo=testrepo&branch=bfdde95432b3af879af969bd2377dc3e55ee46e6
- http://localhost:8002/commits?repo=testrepo&branch=bfdde95432b3af879af969bd2377dc3e55ee46e6&start=fa1330719893098ae397356e8125c2aa45b49221
- http://localhost:8002/commit?repo=testrepo&commit=304c53c163aedfd0c0e0933776f09c24b87f5944

Reviewed at https://reviews.reviewboard.org/r/6972/