Port rb-gateway to go-git.
The libgit2-based git2go is a decent API, but it's a pain to deal with.
It needs very specific versions of libgit2 to be installed, and makes it
very hard to cross-compile or distribute binaries that link against it.
This change ports rb-gateway to the `go-git` library, which is a pure Go
implementation with a very similar API to git2go.

Testing done:
- Ran unit tests
- Tested post-commit review request creation and diff viewing with
  Review Board.

Reviewed at https://reviews.reviewboard.org/r/9875/