Install Mercurial changegroup hooks to trigger post events
RB Gateway is now capable of installing post-receive hooks in a
Mercurial repository that will execute `rb-gateway trigger-webhooks`
with the appropriate arguments to cause a `push` event to occur.

Some of the integration testing infrastructure has been refactored to be
SCM-neutral. Additionally, a bug was fixed in `helpers.CreateTestRepo`
that always recorded the repository as a Git repository.

Testing Done:
- Ran unit tests.
- Ran integration tests.
- Ran `go fmt ./...`.

Reviewed at https://reviews.reviewboard.org/r/10041/