Add a VERSION file and fix the RB Gateway docs name.
This introduces a VERSION file that will be used for the RB Gateway
documentation and packaging process. It simply contains the version in a
standard format parsable by Python and beanbag-tools (allowing for
versions like "2.0beta1").

It also fixes the RB Gateway documentation to not have "documentation"
appear twice in the generated name.

Testing Done:
Built the docs and checked the resulting version and doc names.

Reviewed at https://reviews.reviewboard.org/r/10065/