Move the installation and configuration docs up a level.
When deployed on the website, most of the RB Gateway documentation was
nested in a `rbgateway/<version>/rb-gateway/`, which wasn't really
necessary for namespacing. This change fixes that by moving the docs up
a level in the tree.

Testing Done:
Built the docs. Tested that the links worked, there were no errors, and
the URLs were what I expected.

Reviewed at https://reviews.reviewboard.org/r/10066/